refactor: inline wrestler tag team memberships
